LSU: change lsu structure from pipelined to non pipelined
feedback removed from lsu from previous architecture,
the low number of pins are a result of this. As we don't need to check
for status of the pins. Also the response handling of the memory read
and write operations have been left out. The changes are due to time
constraints.