PC: ability to add immediate and increment the pc given to branch module
unlike PC having a different alu, which runs independently of the system
alu(specific to increment and add immediate offsets on branch
instruction). The branch module is more likely to have this control.
Additional control signals are required to make PC calculate offset and
change its value on the fly. Instead it working as controlled device from
the control unit which provides the value of current PC to the EXU, the
BR module from the exu is able to intercept and forward the new PC back
to the PC.
Since previous implementation also required 32 bit input and output bus
to transfer PC and retrieve Immediate. This is retained in the new
implementation but the additional control circuit is removed.
Also comparison instruction can be issued along with the branch from the
CU. this makes it possible to compare and calculate offset bound to exu.